+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0
 
 * Migrates package and tests to 'platform:web'.
+* Fixes infinite event loop caused by `seekTo` when the video ends.
 
 ## 2.2.0